+++ src/main/java/view/modelCardControllers/ProductCardController.java
@@ -1,5 +1,6 @@
 package view.modelCardControllers;
 
+import dataAccess.OrderDAO;
 import dataAccess.ProductDAO;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -7,15 +8,20 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
+import javafx.scene.paint.Color;
 import javafx.stage.Stage;
+import model.Order_;
 import model.Product;
 import view.Application;
+import view.MessagePrinter;
 
 import java.awt.event.ActionEvent;
 import java.io.IOException;
+import java.util.List;
 
+/**
+ * Controller class for managing product information displayed on a card in the UI.
+ */
 public class ProductCardController {
 
     @FXML
@@ -27,7 +33,6 @@ public class ProductCardController {
     @FXML
     private Label productPrice;
 
-
     @FXML
     private TextField editNameTextFiled;
     @FXML
@@ -36,18 +41,26 @@ public class ProductCardController {
     private TextField editQuantityTextFiled;
     @FXML
     private TextField productIdTextField;
+    @FXML
+    private Label popupLabel;
 
     ProductDAO productDAO = new ProductDAO();
+    OrderDAO orderDAO = new OrderDAO();
 
-
+    /**
+     * Sets the data of the product in the UI.
+     * @param product The product object containing the data to be displayed.
+     */
     public void setData(Product product){
-
         productId.setText(String.valueOf(product.getProductId()));
         productName.setText(product.getName());
         productQuantity.setText(String.valueOf(product.getQuantity()));
         productPrice.setText(String.valueOf(product.getPrice()));
     }
 
+    /**
+     * Retrieves product data and populates the edit view.
+     */
     public void retrieveProductData(){
         editNameTextFiled.setText(productName.getText());
         editPriceTextField.setText(productPrice.getText());
@@ -55,6 +68,10 @@ public class ProductCardController {
         productIdTextField.setText(productId.getText());
     }
 
+    /**
+     * Opens a new window for editing product information.
+     * @throws IOException If an IO error occurs.
+     */
     public void editProduct() throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(Application.class.getResource("pagesContents/editProductPane.fxml"));
         Parent root = fxmlLoader.load();
@@ -70,20 +87,44 @@ public class ProductCardController {
         retrieveProductData();
     }
 
-
+    /**
+     * Updates product information in the database based on changes made in the edit view.
+     */
     public void editProductInDataBase(){
         Product product = productDAO.findById(Integer.parseInt(productIdTextField.getText()));
-        System.out.println(product.getName());
-
         product.setPrice(Float.valueOf(editPriceTextField.getText()));
         product.setQuantity(Integer.valueOf(editQuantityTextFiled.getText()));
         product.setName(editNameTextFiled.getText());
-
         productDAO.update(product);
     }
 
+    /**
+     * Checks if the product is associated with any order.
+     * @param productId The ID of the product to check.
+     * @return True if the product is associated with an order, otherwise false.
+     */
+    public boolean containsProduct(Integer productId){
+        List<Order_> ordersList = orderDAO.findAll();
+        for(Order_ order : ordersList){
+            if(order.getProductId().equals(productId)){
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Deletes the product from the database.
+     */
     public void deleteProductFromDataBase(){
         Product product = productDAO.findById(Integer.parseInt(productIdTextField.getText()));
-        productDAO.delete(product);
+        if(containsProduct(product.getProductId())){
+            popupLabel.setTextFill(Color.RED);
+            new MessagePrinter().showMessage("The product is associated with an order", popupLabel);
+        }
+        else {
+            productDAO.delete(product);
+            new MessagePrinter().showMessage("Product deleted successfully", popupLabel);
+        }
     }
 }
